requirements.txt
fastapi
uvicorn
pymongo
PyJWT
pydub
requests
python-multipart
python-jose